refactor: lógica de recuperação de senha passada para o contexto de usuário
