+++ src/contexts/user/index.tsx
@@ -1,6 +1,6 @@
 import { createContext, useContext, useState } from "react";
 import * as T from "./types";
-import { IUser, IUserLoginRequest } from "@/types/user";
+import { IUser, IUserLoginRequest, IUserRecoverRequest } from "@/types/user";
 import { useRouter } from "next/navigation";
 import { toast } from "react-toast";
 
@@ -45,7 +45,39 @@ export default function UserProvider({ children }: T.IUserProvider) {
   };
 
   const register = () => {};
-  const recover = () => {};
+
+  const recover = (data: IUserRecoverRequest) => {
+    try {
+      const userExists = database.find((user) => user.email === data.email);
+
+      if (!userExists) {
+        throw new Error("Usuário não encontrado");
+      }
+
+      const newDatabase = database.map((user) => {
+        if (user.email === data.email) {
+          user.password = data.newPassword;
+          return user;
+        } else {
+          return user;
+        }
+      });
+
+      setDatabase(newDatabase);
+
+      toast.success("Usuário atualizado com sucesso!");
+
+      setTimeout(() => {
+        toast.success("Você será redirecionado para a tela de login");
+      }, 2000);
+
+      setTimeout(() => {
+        router.push("/");
+      }, 3500);
+    } catch (error: any) {
+      toast.error(error.message);
+    }
+  };
 
   return (
     <userContext.Provider value={{ login, register, recover }}>
